refactor(website): Reorganizes playground components
