+++ website/src/components/FormMappingPlayground/SchemaSelector.tsx
@@ -1,0 +1,30 @@
+import React from "react";
+import { schemas } from "./schemas";
+import { styles } from "./styles";
+
+interface SchemaSelectorProps {
+  label: string;
+  value: string;
+  onChange: (value: string) => void;
+}
+
+export const SchemaSelector: React.FC<SchemaSelectorProps> = ({
+  label,
+  value,
+  onChange,
+}) => (
+  <div style={{ flex: 1 }}>
+    <label style={styles.label}>{label}</label>
+    <select
+      style={styles.select}
+      value={value}
+      onChange={(e) => onChange(e.target.value)}
+    >
+      {schemas.map((opt) => (
+        <option key={opt.id} value={opt.id}>
+          {opt.label}
+        </option>
+      ))}
+    </select>
+  </div>
+);
