+++ website/src/components/FormMappingPlayground/styles.ts
@@ -1,0 +1,66 @@
+export const styles = {
+  container: {
+    maxWidth: 1200,
+    margin: "2rem auto",
+    padding: "2rem",
+    borderRadius: "1rem",
+    background: "#fff",
+    boxShadow: "0 2px 16px rgba(0,0,0,0.08)",
+    fontFamily: "system-ui, sans-serif",
+    display: "flex",
+    flexDirection: "column" as const,
+    gap: "1.5rem",
+  },
+  formContainer: {
+    display: "flex",
+    gap: "2rem",
+  },
+  formSection: {
+    flex: 1,
+    padding: "1rem",
+    borderRadius: "0.5rem",
+    background: "#f8f9fa",
+  },
+  formHeader: {
+    marginBottom: "1rem",
+    fontSize: "1.2rem",
+    fontWeight: 600,
+    color: "#333",
+  },
+  row: {
+    display: "flex",
+    gap: "1rem",
+  },
+  label: {
+    fontWeight: 600,
+    marginBottom: 4,
+    display: "block",
+  },
+  select: {
+    padding: "0.5rem",
+    borderRadius: 6,
+    border: "1px solid #ccc",
+    fontSize: 16,
+    width: "100%",
+  },
+  button: {
+    padding: "0.75rem 1.5rem",
+    borderRadius: 6,
+    border: "none",
+    background: "#1976d2",
+    color: "#fff",
+    fontWeight: 700,
+    fontSize: 16,
+    cursor: "pointer",
+    alignSelf: "flex-start",
+  },
+  code: {
+    background: "#f5f5f5",
+    borderRadius: 6,
+    padding: "1rem",
+    fontFamily: "monospace",
+    fontSize: 14,
+    whiteSpace: "pre-wrap" as const,
+    wordBreak: "break-all" as const,
+  },
+};
