+++ website/src/components/FormMappingPlayground/types.ts
@@ -1,0 +1,9 @@
+export type JsonValue =
+  | string
+  | number
+  | boolean
+  | null
+  | JsonValue[]
+  | { [key: string]: JsonValue };
+
+export type FormData = Record<string, JsonValue>;
