+++ website/src/pages/protocol/playground.astro
@@ -1,5 +1,5 @@
 ---
-import FormMappingPlayground from "../../components/FormMappingPlayground.tsx";
+import FormMappingPlayground from "../../components/FormMappingPlayground";
 ---
 
 <FormMappingPlayground client:load />
